Lab_1/requirements.txt
black==22.12.0
click==8.1.3
glfw==2.5.5
multipledispatch==0.6.0
mypy-extensions==0.4.3
numpy==1.23.5
pathspec==0.10.3
Pillow==9.3.0
platformdirs==2.6.0
PyOpenGL==3.1.6
pyrr==0.10.3
six==1.16.0
tomli==2.0.1
